Minor rework.
